Extract KeySet and MapKey out of Cache
Because they probably going to be shared with ManagedCache
